+++ st2actions/st2actions/container/base.py
@@ -121,7 +121,7 @@ class RunnerContainer(object):
             # mark execution as failed.
             status = action_constants.LIVEACTION_STATUS_FAILED
             # include the error message and traceback to try and provide some hints.
-            result = {'message': str(ex), 'traceback': ''.join(traceback.format_tb(tb, 20))}
+            result = {'error': str(ex), 'traceback': ''.join(traceback.format_tb(tb, 20))}
             context = None
         finally:
             # Log action completion